pyproject.toml
[tool.poetry]
name = "mecab-function"
version = "0.1.0"
description = ""
authors = ["YABATANIEN, Murichaduke <mail@example.com>"]

[tool.poetry.dependencies]
python = "^3.7"
mecab-python3 = "~=1.0.0"
unidic-lite = "~=1.0.0"

[tool.poetry.dev-dependencies]
tox = "^3.24.3"
flake8 = "^3.9.2"
mypy = "^0.910"
black = "^21.8b0"
isort = "^5.9.3"
flake8-import-order = "^0.18.1"

[tool.black]
line-length = 100
exclude = '''
(
    node_modules
    | .mypy_cache
    | .pytest_cache
    | .tox
    | venv
)
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
